+++ project-base/app/.env
@@ -56,10 +56,10 @@ AKENEO_SECRET=''
 AKENEO_USER=''
 AKENEO_PASSWORD=''
 
-SSFWCC_BRIDGE_ENABLED=false
-SSFWCC_BRIDGE_BASE_URI='https://api.example.com/api/'
-SSFWCC_BRIDGE_USER=''
-SSFWCC_BRIDGE_PASSWORD=''
+DATA_BRIDGE_ENABLED=false
+DATA_BRIDGE_BASE_URI='https://api.example.com/api/'
+DATA_BRIDGE_USER=''
+DATA_BRIDGE_PASSWORD=''
 
 LANGUAGE_CONSTANTS_URL_PATTERN='http://webserver:8080/locales/%s/common.json'
