+++ project-base/app/deploy/deploy-project.sh
@@ -46,10 +46,10 @@ function deploy() {
         ["AKENEO_USER"]=${AKENEO_USER}
         ["AKENEO_PASSWORD"]=${AKENEO_PASSWORD}
 
-        ["SSFWCC_BRIDGE_ENABLED"]=${SSFWCC_BRIDGE_ENABLED}
-        ["SSFWCC_BRIDGE_BASE_URI"]=${SSFWCC_BRIDGE_BASE_URI}
-        ["SSFWCC_BRIDGE_USER"]=${SSFWCC_BRIDGE_USER}
-        ["SSFWCC_BRIDGE_PASSWORD"]=${SSFWCC_BRIDGE_PASSWORD}
+        ["DATA_BRIDGE_ENABLED"]=${DATA_BRIDGE_ENABLED}
+        ["DATA_BRIDGE_BASE_URI"]=${DATA_BRIDGE_BASE_URI}
+        ["DATA_BRIDGE_USER"]=${DATA_BRIDGE_USER}
+        ["DATA_BRIDGE_PASSWORD"]=${DATA_BRIDGE_PASSWORD}
 
         ["S3_ENDPOINT"]=${S3_ENDPOINT}
         ["S3_ACCESS_KEY"]=${S3_ACCESS_KEY}
