+++ project-base/app/src/Component/DataBridge/Transfer/DummyImportTransferFacade.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Component\SsfwccBridge\Transfer;
+namespace App\Component\DataBridge\Transfer;
 
 use Generator;
