+++ project-base/app/src/Component/DataBridge/Transfer/Exception/TransferException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Component\SsfwccBridge\Transfer\Exception;
+namespace App\Component\DataBridge\Transfer\Exception;
 
 use Exception;
