+++ project-base/app/src/Component/DataBridge/Transfer/Exception/TransferInvalidDataAdministratorCriticalException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Component\SsfwccBridge\Transfer\Exception;
+namespace App\Component\DataBridge\Transfer\Exception;
 
 class TransferInvalidDataAdministratorCriticalException extends TransferInvalidDataException
 {
